prototype for adding Prometheus service discovery
I think it could be useful for others having a Prometheus Service Discovery endpoint.
The idea behind is having an output ready to be used for Prometheus (https://www.robustperception.io/using-json-file-service-discovery-with-prometheus/); in such a way if the cluster has dynamic members joining, you can have them monitored with no human intervention. All you need is periodically poll this exporter ( on all masters), save on file if the output is not 'null' and configure Prometheus to reads this file. 
THIS IS A PROTOTYPE to start a discussion about this enhancement, 
it needs to modified in order to :
* written in nice Golang :)
* use any json library instead of creating a string 
* use command line arguments instead of Environment variables
* decide what to do on a slave
* have some testing code
* have the related code outside the main file